Cleanup for ComponentTemplates When doing bulk template registration only process one template at a time. If we need to handle backup it could break running multiple. DiscordCommand cleanup Updated Embed and Embed Field templates to process their types in their class Added DiscordGuild.GetChannel(Snowflake) to return a guild channel or thread channel Removed Obsolete From DiscordInteraction.GetLangMessage & DiscordLang.GetDiscordInteractionLangMessage Made DiscordLang a proper Library Added DiscordInteraction & DiscordMessage placeholders Added DiscordPool.GetStringBuilder(string initial) Added DiscordUser.IsBot & DiscordUser.IsSystem properties Added GuildMember.IsBot property PermissionFlagsStringConverter removed unnecessary unbox Added PlaceholderData.AddMessage(DiscordMessage) Added PlaceholderData.AddInteraction(DiscordInteraction) Added PlaceholderData.AddNowTimestamp() Improved PlaceholderData null checking PlayerExt.SendDiscordTemplateMessage removed language parameter since we can get it from oxide lang

+++ Oxide.Ext.Discord/Json/Converters/PermissionFlagsStringConverter.cs
@@ -23,7 +23,7 @@ namespace Oxide.Ext.Discord.Json.Converters
                 return;
             }
             
-            writer.WriteValue(((ulong)value).ToString());
+            writer.WriteValue(value.ToString());
         }
 
         /// <summary>
